requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
requests
PyPDF2
chromadb
